Restructure tutorial folder names to include numeric sequence
