+++ apps/nextjs-mornington/pages/00_store_tutorial.tsx
@@ -1,3 +1,3 @@
 import React from "react";
-import { SimpleGame } from "../src/tutorial/store/app";
+import { SimpleGame } from "../src/tutorial/00_store/app";
 export default () => <SimpleGame />;
