+++ apps/nextjs-mornington/pages/01_action_tutorial.tsx
@@ -1,3 +1,3 @@
 import React from "react";
-import { GameRoot } from "../src/tutorial/action/app";
+import { GameRoot } from "../src/tutorial/01_action/app";
 export default () => <GameRoot />;
