+++ apps/nextjs-mornington/pages/02_event_tutorial.tsx
@@ -1,3 +1,3 @@
 import React from "react";
-import { GameRoot } from "../src/tutorial/event/app";
+import { GameRoot } from "../src/tutorial/02_event/app";
 export default () => <GameRoot />;
