+++ apps/nextjs-mornington/pages/index.tsx
@@ -4,13 +4,13 @@ export default () => (
     <h1>Example 'apps'</h1>
     <p>The apps linked below are reference examples from the Lauf tutorial</p>
     <p>
-      <a href="./store_tutorial">Store Example</a>
+      <a href="./00_store_tutorial">Store Example</a>
     </p>
     <p>
-      <a href="./action_tutorial">Actions Example</a>
+      <a href="./01_action_tutorial">Actions Example</a>
     </p>
     <p>
-      <a href="./event_tutorial">Event Example</a>
+      <a href="./02_event_tutorial">Event Example</a>
     </p>
   </>
 );
